depository/requirements.txt
django
packaging
pillow
django-filter
django-rest-framework
